make prompts more interactive for dnf hacks, rpm repo, system upgrade, multimedia package installation and portainer install

+++ linux/fedora/post-install/post-install.sh
@@ -1,46 +1,114 @@
 #!/bin/bash
 
 
-# check for these lines in /etc/dnf/dnf.conf
-if grep -Fxq "fastestmirror=True" /etc/dnf/dnf.conf && \
-   grep -Fxq "max_parallel_downloads=10" /etc/dnf/dnf.conf && \
-   grep -Fxq "defaultyes=True" /etc/dnf/dnf.conf; then
-   echo "The lines already exist in the file."
+# Notify user about checking for required lines in /etc/dnf/dnf.conf
+echo "This script will check if the following lines exist in /etc/dnf/dnf.conf:"
+echo "fastestmirror=True"
+echo "max_parallel_downloads=10"
+echo "defaultyes=True"
+
+# Ask user if they want to proceed with checking for the lines
+read -p "Do you want to proceed with checking? [Y/n] " response
+
+# Convert response to uppercase for consistency
+response=${response^^}
+# Check if the user wants to proceed with checking
+if [[ $response == "N" ]]; then
+  echo "Skipping checking for lines in /etc/dnf/dnf.conf."
 else
-  # append the three lines to the file if not already
-  sudo tee -a /etc/dnf/dnf.conf >/dev/null <<EOF
-  fastestmirror=True
-  max_parallel_downloads=10
-  defaultyes=True
+  # Check if the lines exist in /etc/dnf/dnf.conf
+  if grep -Fxq "fastestmirror=True" /etc/dnf/dnf.conf && \
+     grep -Fxq "max_parallel_downloads=10" /etc/dnf/dnf.conf && \
+     grep -Fxq "defaultyes=True" /etc/dnf/dnf.conf; then
+       echo "The lines already exist in the file."
+  else
+    # Append the three lines to the file if not already
+    sudo tee -a /etc/dnf/dnf.conf >/dev/null <<EOF
+    fastestmirror=True
+    max_parallel_downloads=10
+    defaultyes=True
 EOF
     echo "The lines have been appended to the file."
+  fi
 fi
 
 
-# Enable RPM Fusion repositories if not already enabled
-if ! dnf list installed | grep -q 'rpmfusion-free-release'; then
-  sudo dnf install -y https://download1.rpmfusion.org/free/fedora/rpmfusion-free-release-$(rpm -E %fedora).noarch.rpm
+
+# Prompt user to enable RPM Fusion repositories
+echo "This script will enable RPM Fusion free and non-free repositories."
+read -p "Do you want to enable both free and non-free repositories? [Y/n] " response
+
+# Convert response to uppercase for consistency
+response=${response^^}
+if [[ $response != "N" ]]; then
+    # Enable RPM Fusion Free repositories if not already enabled
+    if ! dnf list installed | grep -q 'rpmfusion-free-release'; then
+        sudo dnf install -y https://download1.rpmfusion.org/free/fedora/rpmfusion-free-release-$(rpm -E %fedora).noarch.rpm
+    fi
+
+    # Enable RPM Fusion NonFree repositories if not already enabled
+    if ! dnf list installed | grep -q 'rpmfusion-nonfree-release'; then
+        sudo dnf install -y https://download1.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-$(rpm -E %fedora).noarch.rpm
+    fi
+
+    echo "RPM Fusion free and non-free repositories have been enabled."
+
+    # Ask the user if they want to update the core group
+    read -p "Do you want to update the core group now? [Y/n] " response
+
+    # Convert response to uppercase for consistency
+    response=${response^^}
+    if [[ $response != "N" ]]; then
+        # Update the core group
+        sudo dnf group update core
+        echo "Group core updated successfully!"
+    else
+        echo "Skipping updating of the core group."
+    fi
+else
+    echo "Skipping enabling of RPM Fusion repositories."
 fi
 
-if ! dnf list installed | grep -q 'rpmfusion-nonfree-release'; then
-  sudo dnf install -y https://download1.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-$(rpm -E %fedora).noarch.rpm
-fi
 
-sudo dnf group update core
 
+# Define the required multimedia packages and the LAME library
+multimedia_packages=("gstreamer1-plugins-bad-free" "gstreamer1-plugins-good" "gstreamer1-plugins-base" "gstreamer1-plugin-openh264" "gstreamer1-libav")
+lame_library="lame"
 
 # Check if required multimedia packages are already installed
-if ! dnf list installed | grep -q gstreamer1-plugins-bad-free; then
-    # Install plugins for playing movies and music
-    sudo dnf install gstreamer1-plugins-{bad-\*,good-\*,base} gstreamer1-plugin-openh264 gstreamer1-libav --exclude=gstreamer1-plugins-bad-free-devel
+if ! dnf list installed "${multimedia_packages[@]}" &> /dev/null; then
+    # Prompt the user to install the multimedia packages
+    read -p "You need to install some multimedia packages to play music and videos on your system. Do you want to install them now? [Y/n] " response
+    response=${response^^}
+    if [[ $response == "N" ]]; then
+        echo "Skipping installation of multimedia packages."
+    else
+        sudo dnf install "${multimedia_packages[@]/#/gstreamer1-plugins-}" --exclude=gstreamer1-plugins-bad-free-devel
+    fi
 fi
 
-if ! dnf list installed | grep -q lame; then
-    sudo dnf install lame\* --exclude=lame-devel
+# Check if LAME library is already installed
+if ! dnf list installed "$lame_library" &> /dev/null; then
+    # Prompt the user to install the LAME library
+    read -p "You need to install LAME library to encode and decode MP3 files. Do you want to install it now? [Y/n] " response
+    response=${response^^}
+    if [[ $response == "N" ]]; then
+        echo "Skipping installation of LAME library."
+    else
+        sudo dnf install "$lame_library"* --exclude=lame-devel
+    fi
+
+    # Prompt the user to update the multimedia group
+    read -p "Do you want to update the multimedia group now? [Y/n] " response
+    response=${response^^}
+    if [[ $response == "N" ]]; then
+        echo "Skipping updating of the multimedia group."
+    else
+        sudo dnf group update multimedia
+        echo "Group multimedia updated successfully!"
+    fi
 fi
 
-# Upgrade multimedia packages
-sudo dnf group upgrade --with-optional Multimedia
 
 
 # Declare a variable with the list of basic tools to install
@@ -56,9 +124,18 @@ do
         while true; do
             read -p "$tool is not installed. Do you want to install it? (y/n): " yn
             case $yn in
-                [Yy]* ) sudo dnf install -y "$tool"; break;;
-                [Nn]* ) break;;
-                * ) echo "Please answer y or n.";;
+                [Yy]* )
+                    sudo dnf install -y "$tool"
+                    echo "$tool has been installed."
+                    break
+                    ;;
+                [Nn]* )
+                    echo "Skipping $tool installation."
+                    break
+                    ;;
+                * )
+                    echo "Please answer y or n."
+                    ;;
             esac
         done
     else
@@ -67,14 +144,42 @@ do
 done
 
 
-# upgrade the system
-sudo dnf --refresh upgrade -y
 
+# Ask the user if they want to proceed with the upgrade
+read -p "Do you want to upgrade your system now? [Y/n] " response
 
-# Check if flathub already enabled
-if ! flatpak remotes | grep -q "flathub"; then
-    flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
+# Convert response to uppercase for consistency
+response=${response^^}
+if [[ $response == "N" ]]; then
+    echo "Skipping system upgrade."
+else
+    # Perform the system upgrade
+    echo "Upgrading system..."
+    sudo dnf --refresh upgrade
+    echo "System upgrade complete!"
 fi
+
+
+
+# Check if Flathub is already enabled
+if flatpak remotes | grep -q "flathub"; then
+    echo "Flathub is already enabled."
+else
+    # Ask the user if they want to enable Flathub
+    read -p "Flathub is not enabled. Do you want to enable it now? [Y/n] " response
+
+    # Convert response to uppercase for consistency
+    response=${response^^}
+
+    if [[ $response == "N" ]]; then
+        echo "Skipping enabling of Flathub."
+    else
+        flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
+        echo "Flathub has been enabled."
+    fi
+fi
+
+
 
 # define the list of flatpak apps to install
 flatpak_apps=("org.kde.kdenlive" "com.github.tchx84.Flatseal" "com.belmoussaoui.Obfuscate" "com.visualstudio.code" "com.bitwarden.desktop")
@@ -107,6 +212,7 @@ do
 done
 
 
+
 # Define the list of software to install
 software=("Thunderbird" "Nextcloud Desktop" "GIMP" "VLC" "Steam")
 
@@ -136,13 +242,28 @@ do
 done
 
 
-# Check if portmaster is already installed
+
+# Check if Portmaster is already installed
 if [ -d "/opt/portmaster" ]; then
     echo "Portmaster is already installed."
 else
-    wget -c --directory-prefix=/tmp/ https://updates.safing.io/latest/linux_amd64/packages/portmaster-installer.rpm
-    sudo dnf install -y /tmp/portmaster-installer.rpm
+    # Ask the user if they want to install Portmaster
+    read -p "Portmaster is not installed. Do you want to install it now? [Y/n] " response
+
+    # Convert response to uppercase for consistency
+    response=${response^^}
+    if [[ $response == "N" ]]; then
+        echo "Skipping installation of Portmaster."
+    else
+        # Download and install Portmaster
+        echo "Downloading Portmaster installer..."
+        wget -c --directory-prefix=/tmp/ https://updates.safing.io/latest/linux_amd64/packages/portmaster-installer.rpm
+        echo "Installing Portmaster..."
+        sudo dnf install -y /tmp/portmaster-installer.rpm
+        echo "Portmaster installed successfully!"
+    fi
 fi
+
 
 
 # Define the list of development tools to install
